requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==2.0.6
django-rest-registration==0.5.5
djangorestframework==3.8.2
ete3==3.1.1
idna==2.9
numpy==1.18.4
packaging==20.3
Pillow==7.1.2
pyparsing==2.4.7
PyQt-builder==1.3.2
PyQt5==5.13.0
PyQt5-sip==12.7.2
pytz==2019.3
requests==2.23.0
sip==5.2.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typing==3.7.4.1
urllib3==1.25.8
whitenoise==5.0.1
bio==0.1.0
biopython==1.76
pandas==1.0.3
python-dateutil==2.8.1
toml==0.10.0
urllib3==1.25.8
wget==3.2
xlrd==1.2.0